C type cleanups

+++ src/duk_hbuffer_ops.c
@@ -189,7 +189,8 @@ size_t duk_hbuffer_insert_xutf8(duk_hthread *thr, duk_hbuffer_dynamic *buf, size
 
 	/* Intentionally no fast path: insertion is not that central */
 
-	len = duk_unicode_encode_xutf8(codepoint, tmp);
+	/* FIXME: cp -> duk_codepoint_t */
+	len = (size_t) duk_unicode_encode_xutf8(codepoint, tmp);
 	duk_hbuffer_insert_bytes(thr, buf, offset, tmp, len);
 	return len;
 }
@@ -211,7 +212,8 @@ size_t duk_hbuffer_insert_cesu8(duk_hthread *thr, duk_hbuffer_dynamic *buf, size
 
 	/* Intentionally no fast path: insertion is not that central */
 
-	len = duk_unicode_encode_cesu8(codepoint, tmp);
+	/* FIXME: cp -> duk_codepoint_t */
+	len = (size_t) duk_unicode_encode_cesu8(codepoint, tmp);
 	duk_hbuffer_insert_bytes(thr, buf, offset, tmp, len);
 	return len;
 }
@@ -292,7 +294,8 @@ size_t duk_hbuffer_append_xutf8(duk_hthread *thr, duk_hbuffer_dynamic *buf, duk_
 		return 1;
 	}
 
-	len = duk_unicode_encode_xutf8(codepoint, tmp);
+	/* FIXME: cp -> duk_codepoint_t */
+	len = (size_t) duk_unicode_encode_xutf8(codepoint, tmp);
 	duk_hbuffer_insert_bytes(thr, buf, DUK_HBUFFER_GET_SIZE(buf), tmp, len);
 	return len;
 }
@@ -320,7 +323,8 @@ size_t duk_hbuffer_append_cesu8(duk_hthread *thr, duk_hbuffer_dynamic *buf, duk_
 		return 1;
 	}
 
-	len = duk_unicode_encode_cesu8(codepoint, tmp);
+	/* FIXME: cp -> duk_codepoint_t */
+	len = (size_t) duk_unicode_encode_cesu8(codepoint, tmp);
 	duk_hbuffer_insert_bytes(thr, buf, DUK_HBUFFER_GET_SIZE(buf), tmp, len);
 	return len;
 }
